+++ src/problems/vrptw/vrptw.cpp
@@ -168,7 +168,8 @@ Solution VRPTW::solve(unsigned exploration_level,
   assert(nb_init_solutions <= parameters.size());
 
   std::vector<TWSolution> tw_solutions(nb_init_solutions);
-  std::vector<utils::SolutionIndicators> sol_indicators(nb_init_solutions);
+  std::vector<utils::SolutionIndicators<TWRoute>> sol_indicators(
+    nb_init_solutions);
 #ifdef LOG_LS_OPERATORS
   std::vector<std::array<ls::OperatorStats, OperatorName::MAX>> ls_stats(
     nb_init_solutions);